feat: add unsquashfs util to Squashfs
This is mostly a copy from `backhand-cli`'s `unsquashfs`, but with all
the args/progress handling removed and more thorough error handling.

Due to the dependency on `rayon`, the utility is currently gated behind
the `util` feature of `backhand`.

I tried also gating the `nix` dependency on the feature, but it was hard
due to it showing up in `BackhandError`.

Fixes: #354